chore: Eliminate duplicate upgrade code
